API/requirements.txt
requests~=2.28.2
fastapi~=0.95.0
uvicorn~=0.21.0
pydantic~=1.10.6
pymongo~=4.3.3
paho-mqtt~=1.6.1